% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_stadia_api_key}
\alias{set_stadia_api_key}
\title{Set Stadia Maps API key (environment variable)}
\usage{
set_stadia_api_key(key)
}
\arguments{
\item{key}{Your Stadia Maps API key (string)}
}
\description{
Set Stadia Maps API key (environment variable)
}
