% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{stadia_get_tile}
\alias{stadia_get_tile}
\title{Download a raster tile (z/x/y) as raw PNG}
\usage{
stadia_get_tile(z, x, y, style = "outdoors", file = NULL)
}
\arguments{
\item{z}{zoom level (integer)}

\item{x}{tile X (integer)}

\item{y}{tile Y (integer)}

\item{style}{map style; Stadia supports 'outdoors'/'alidade_smooth' etc. Default 'outdoors'}

\item{file}{(optional) file path to save the tile image to}
}
\value{
raw vector (PNG) invisibly or writes file if path provided
}
\description{
Download a raster tile (z/x/y) as raw PNG
}
