% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{stadia_plot_png}
\alias{stadia_plot_png}
\title{Plot a PNG image returned by stadia_read_png}
\usage{
stadia_plot_png(img)
}
\arguments{
\item{img}{image to plot}
}
\description{
Plot a PNG image returned by stadia_read_png
}
