% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{stadia_read_png}
\alias{stadia_read_png}
\title{Read PNG bytes into R as a raster image for plotting}
\usage{
stadia_read_png(raw)
}
\arguments{
\item{raw}{raw vector from stadia_get_tile or stadia_static_map}
}
\value{
raster array (matrix-like) for plotting with rasterImage
}
\description{
Read PNG bytes into R as a raster image for plotting
}
